Broaden Map component to handle a more generic class called "area" instead of just room.
Added regions to possible area types and propagated room type functionality to all areas including the new regions.  Because of the difference in schema, there is still a lot of selection logic occuring in the mouse and key handlers but if it becomes unwieldy I can refactor.

Updated styles for area overlays and changed zebes overlays to a light color.

Implemented grid regions for the brinstar map.