Add textures and hover interactions to the main toolbar actions.
Keep track of the current sub-toolbar in the Toolbars component so that a visual cue can be maintained to show the user what bar is active.

Change the actions available in the overworld.  Will be working on a shop sub-toolbar.

A lot of styles were duplicated from Overlay to Toolbars.  Refactor eventually.