Turned coop-client.js into a Svelte writable store and enabled it.
Every call assumes the caller doesn't know whether a connection is active.  The _modeEnabled state stores whether a session is active, and the tracker just treats it as a subscriber to map data (like storage and map stats).

Should allow for robust tracking whether or not the server is available.  Will implement some type of toast notification on the tracker page.